+++ programs/zoomi-smart-contract/src/state/scooter.rs
@@ -6,7 +6,7 @@ pub struct Scooter {
     pub id: u32,                        // Scooter unique ID from DB
     pub shopkeeper_id: u32,             // Shopkeeper unique ID from DB
     pub zoomi_device_pubkey: Pubkey,    // Zoomi device public key associated with the scooter
-    pub hourly_rate: u16,               // Price per hour in USDC
+    pub hourly_rate: u64,               // Price per hour in USDC
     pub status: ScooterStatus,          // Scooter status
     pub location_lat: i32,              // i32 = f32 * 1000000
     pub location_long: i32,             // i32 = f32 * 1000000
